server-a-job-analysis/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
# MCP SDK for true MCP protocol
mcp>=0.9.0
openai>=1.50.0
python-dotenv>=1.0.0
